segorang-server/requirements.txt
boto3==1.20.37
botocore==1.23.37
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Faker==11.3.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-validation-extended==0.1.7
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
loguru==0.6.0
MarkupSafe==2.0.1
Pillow==9.0.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
text-unidecode==1.3
urllib3==1.26.7
Werkzeug==2.0.2
win32-setctime==1.1.0
sejong-univ-auth
pymysql